requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
facebook-sdk==0.4.0
itsdangerous==0.23
protobuf==2.5.0
rethinkdb==1.11.0-1
wsgiref==0.1.2
